+++ examples/vila/run_vila.py
@@ -1,8 +1,8 @@
 
 import torchpipe
-import cv2, torch
+import cv2, torch, os, glob
 
-torchpipe.utils.cpp_extension.load(name="plugin", sources=["./TorchPlugin.cpp"],
+torchpipe.utils.cpp_extension.load(name="plugin", sources=glob.glob("./*.cpp"),
                                    extra_include_paths=['/workspace/TensorRT-10.2.0.19/include/'],
                                    extra_ldflags=['-L/workspace/TensorRT-10.2.0.19/targets/x86_64-linux-gnu/lib/','-lnvinfer_plugin','-lnvinfer'],
                                    verbose=True)
@@ -22,8 +22,11 @@ def run(inputs):
 img_path  = 'demo_images/av.png'
 img = cv2.imread(img_path)
 input = torch.from_numpy(img)
-input = torch.zeros((244,2560))
-data = {'data':input,'node_name':'batchful'}
+input = torch.ones((244,2560)).to(torch.float16).cuda()
+pt = "/workspace/VILA/inputs_embeds.pt"
+if os.path.exists(pt):
+    input = torch.load(pt).squeeze(0).cuda()
+data = {'data':input,'request_id':'zzz', 'node_name':'batchful', 'trt_plugin':'batchless_prefill'}
 model(data)
 print(data['result'].shape)
 # torchpipe.utils.test.test_from_raw_file(run,file_dir="../assets/", num_clients=2,